added def of HIGH_PREC_INPUT as workaround and added new version of gmp lib

+++ Processor/FloatInput.h
@@ -19,7 +19,7 @@ public:
 
     const static int TYPE = 2;
 
-    long items[N_DEST];
+    long long items[N_DEST];
 
     void read(std::istream& in, const int* params);
 };
